+++ inte_screening/models/__init__.py
@@ -1,5 +1,8 @@
 from .daily_closing_log import DailyClosingLog
-from .daily_closing_log_revised import DailyClosingLogRevised
+from .daily_closing_log_revised import (
+    DailyClosingLogRevised,
+    get_daily_log_revision_date,
+)
 from .signals import subject_refusal_on_post_delete, subject_refusal_on_post_save
 from .subject_refusal import SubjectRefusal
 from .subject_screening import SubjectScreening