+++ tests/inte_screening/test_daily_log_revised.py
@@ -1,3 +1,5 @@
+import pdb
+
 from dateutil.relativedelta import relativedelta
 from django.contrib.sites.models import Site
 from django.test import TestCase, override_settings, tag
@@ -26,7 +28,10 @@ class TestDailyLogRevised(InteTestCaseMixin, WebTest):
         }
 
     @tag("daily")
-    @override_settings(SITE_ID=103)
+    @override_settings(
+        SITE_ID=103,
+        INTE_SCREENING_DCL_REVISION_DATE=get_utcnow().date() + relativedelta(days=-10),
+    )
     def test_integrated_for_intervention_site(self):
         form = DailyClosingLogRevisedForm(data=self.data)
         form.is_valid()
@@ -39,14 +44,20 @@ class TestDailyLogRevised(InteTestCaseMixin, WebTest):
         self.assertNotIn("clinic_services", form._errors)
 
     @tag("daily")
-    @override_settings(SITE_ID=101)
+    @override_settings(
+        SITE_ID=101,
+        INTE_SCREENING_DCL_REVISION_DATE=get_utcnow().date() + relativedelta(days=-10),
+    )
     def test_integrated_for_control_site(self):
         form = DailyClosingLogRevisedForm(data=self.data)
         form.is_valid()
         self.assertIn("clinic_services", form._errors)
 
     @tag("daily")
-    @override_settings(SITE_ID=103)
+    @override_settings(
+        SITE_ID=103,
+        INTE_SCREENING_DCL_REVISION_DATE=get_utcnow().date() + relativedelta(days=-10),
+    )
     def test_clinic_start_time(self):
         IntegratedCareClinicRegistration.objects.create(date_opened=get_utcnow().date())
         data = {